backend/requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
alembic>=1.7.1
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
python-multipart>=0.0.5
aiofiles>=0.8.0
jinja2>=3.0.1
requests
